+++ smarts/sstudio/types.py
@@ -24,7 +24,17 @@ import random
 from dataclasses import dataclass, field
 from enum import IntEnum
 from sys import maxsize
-from typing import Any, Callable, Dict, List, Optional, Sequence, Tuple, Union
+from typing import (
+    Any,
+    Callable,
+    Dict,
+    FrozenSet,
+    List,
+    Optional,
+    Sequence,
+    Tuple,
+    Union,
+)
 
 import numpy as np
 from shapely.affinity import rotate as shapely_rotate
@@ -40,6 +50,7 @@ from shapely.geometry import (
 from shapely.ops import split, unary_union
 
 from smarts.core import gen_id
+from smarts.core.colors import Colors
 from smarts.core.coordinates import RefLinePoint
 from smarts.core.default_map_builder import get_road_map
 from smarts.core.road_map import RoadMap
@@ -352,8 +363,8 @@ class MapSpec:
 
     source: str
     """A path or URL or name uniquely designating the map source."""
-    lanepoint_spacing: Optional[float] = None
-    """If specified, the default distance between pre-generated Lane Points (Waypoints)."""
+    lanepoint_spacing: float = 1.0
+    """The default distance between pre-generated Lane Points (Waypoints)."""
     default_lane_width: Optional[float] = None
     """If specified, the default width (in meters) of lanes on this map."""
     shift_to_origin: bool = False
@@ -1063,6 +1074,16 @@ class TrafficHistoryDataset:
 
 
 @dataclass(frozen=True)
+class ScenarioMetadata:
+    """Scenario data that does not have influence on simulation."""
+
+    actor_of_interest_re_filter: str
+    """Vehicles with names that match this pattern are vehicles of interest."""
+    actor_of_interest_color: Colors
+    """The color that the vehicles of interest should have."""
+
+
+@dataclass(frozen=True)
 class Scenario:
     """The sstudio scenario representation."""
 
@@ -1087,3 +1108,5 @@ class Scenario:
     """Friction coefficient of patches of road surface."""
     traffic_histories: Optional[Sequence[Union[TrafficHistoryDataset, str]]] = None
     """Traffic vehicles trajectory dataset to be replayed."""
+    scenario_metadata: Optional[ScenarioMetadata] = None
+    """"Scenario data that does not have influence on simulation."""